+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased](https://github.com/dalance/svls/compare/v0.1.1...Unreleased) - ReleaseDate
 
+* [Added] Show parse error
+* [Changed] Enable all rules when `.svlint.toml` is not found
+
 ## [v0.1.1](https://github.com/dalance/svlint/compare/v0.1.0...v0.1.1) - 2019-11-25
 
 * [Added] Error message about `.svlint.toml` is not found